+++ futures/tests/stream_select_all.rs
